requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.0

# Hugging Face
datasets>=2.16.0
huggingface-hub>=0.20.0

# Data augmentation
albumentations>=1.3.1
opencv-python-headless>=4.8.0

# Metrics & evaluation
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Grad-CAM
pytorch-grad-cam>=1.5.0

# Logging
tensorboard>=2.14.0
tqdm>=4.66.0

# Config
pyyaml>=6.0.1
omegaconf>=2.3.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
Pillow>=10.0.0

# Web demo
gradio>=4.0.0

# Export
onnx>=1.15.0
onnxruntime>=1.16.0

# Report generation
fpdf2>=2.7.0

# Testing
pytest>=7.4.0
